+++ mathics/builtin/numbertheory.py
@@ -754,16 +754,33 @@ class PowerMod(Builtin):
 class Prime(SympyFunction):
     """
     <dl>
-    <dt>'Prime[$n$]'
-        <dd>returns the $n$th prime number.
+      <dt>'Prime[$n$]'
+      <dt>'Prime'[{$n0$, $n1$, ...}]
+      <dd>returns the $n$th prime number where $n$ is an positive Integer.
+      If given a list of integers, the return value is a list with 'Prime' applied to each.
     </dl>
 
+    Note that the first prime is 2, not 1:
     >> Prime[1]
      = 2
 
     >> Prime[167]
      = 991
+
+    When given a list of integers, a list is returned:
+    >> Prime[{5, 10, 15}]
+     = {11, 29, 47}
+
+    1.2 isn't an integer
+    >> Prime[1.2]
+     = Prime[1.2]
+
+    Since 0 is less than 1, like 1.2 it is invalid.
+    >> Prime[{0, 1, 1.2, 3}]
+     = {Prime[0], 2, Prime[1.2], 5}
     """
+
+    attributes = ("Listable", "NumericFunction")
 
     def apply(self, n, evaluation):
         "Prime[n_]"
@@ -774,12 +791,16 @@ class Prime(SympyFunction):
             return SympyPrime(expr.leaves[0].to_sympy(**kwargs))
 
 
-class PrimePi(Builtin):
+class PrimePi(SympyFunction):
     """
     <dl>
     <dt>'PrimePi[$x$]'
         <dd>gives the number of primes less than or equal to $x$.
     </dl>
+
+    PrimePi is the inverse of Prime:
+    >> PrimePi[2]
+     = 1
 
     >> PrimePi[100]
      = 25
@@ -793,6 +814,12 @@ class PrimePi(Builtin):
     >> PrimePi[E]
      = 1
     """
+
+    sympy_name = "ntheory.primepi"
+    mpmath_name = "primepi"
+
+    attributes = ("Listable", "NumericFunction")
+
 
     # TODO: Traditional Form
 
@@ -862,7 +889,7 @@ class PrimePowerQ(Builtin):
             return Symbol("False")
 
 
-class PrimeQ(Builtin):
+class PrimeQ(SympyFunction):
     """
     <dl>
     <dt>'PrimeQ[$n$]'
@@ -896,7 +923,9 @@ class PrimeQ(Builtin):
      = {False, True, True, False, True, False, True, False, False, False, True, False, True, False, False, False, True, False, True, False}
     """
 
-    attributes = ("Listable",)
+    attributes = ("Listable", "NumericFunction")
+
+    sympy_name = "isprime"
 
     def apply(self, n, evaluation):
         "PrimeQ[n_]"